Move tests for collect(_granular) to  `test_stage_load.py`
No other tests have been added.
Some replacements from `dvc.collect` and `dvc.collect_granular`
to `dvc.stage.*` has been made.
